+++ lib/recipes/chemistry/recipe.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-CHEMISTRY_RECIPE = Kitchen::BookRecipe.new(book_short_name: :chemistry) do |doc, resources|
+CHEMISTRY_RECIPE = Kitchen::BookRecipe.new(book_short_name: :chemistry) do |doc, _resources|
   include Kitchen::Directions
 
   book = doc.book
@@ -9,7 +9,6 @@ CHEMISTRY_RECIPE = Kitchen::BookRecipe.new(book_short_name: :chemistry) do |doc,
   # Some stuff just goes away
   book.search('cnx-pi').trash
 
-  BakeImages.v1(book: book, resources: resources)
   BakeUnnumberedFigure.v1(book: book)
   BakePreface.v1(book: book)