+++ lib/recipes/english-composition/recipe.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 ENGLISH_COMPOSITION_RECIPE = Kitchen::BookRecipe.new(book_short_name: :english_composition) \
-do |doc, resources|
+do |doc, _resources|
   include Kitchen::Directions
 
   book = doc.book
@@ -10,7 +10,6 @@ do |doc, resources|
   # Some stuff just goes away
   book.search('cnx-pi').trash
 
-  BakeImages.v1(book: book, resources: resources)
   BakeUnnumberedFigure.v1(book: book)
   BakePreface.v1(book: book)
   BakeHandbook.v1(book: book)