+++ lib/recipes/pl-economics/recipe.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-PL_ECONOMICS_RECIPE = Kitchen::BookRecipe.new(book_short_name: :pleconomics) do |doc, resources|
+PL_ECONOMICS_RECIPE = Kitchen::BookRecipe.new(book_short_name: :pleconomics) do |doc, _resources|
   include Kitchen::Directions
 
   doc.selectors.override(
@@ -14,7 +14,6 @@ PL_ECONOMICS_RECIPE = Kitchen::BookRecipe.new(book_short_name: :pleconomics) do 
   book.search('cnx-pi').trash
   BakeListsWithPara.v1(book: book)
 
-  BakeImages.v1(book: book, resources: resources)
   BakePreface.v1(book: book, title_element: 'h1')
   BakeUnnumberedFigure.v1(book: book)
   BakeChapterIntroductions.v2(book: book, options: { cases: true })