+++ lib/recipes/python/recipe.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-PYTHON_RECIPE = Kitchen::BookRecipe.new(book_short_name: :python) do |doc, resources|
+PYTHON_RECIPE = Kitchen::BookRecipe.new(book_short_name: :python) do |doc, _resources|
   include Kitchen::Directions
 
   book = doc.book
@@ -8,7 +8,6 @@ PYTHON_RECIPE = Kitchen::BookRecipe.new(book_short_name: :python) do |doc, resou
 
   book.search('cnx-pi').trash
 
-  BakeImages.v1(book: book, resources: resources)
   BakePreface.v1(book: book)
   BakeUnnumberedFigure.v1(book: book)
   BakeUnclassifiedNotes.v1(book: book)