+++ lib/recipes/u-physics/recipe.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-U_PHYSICS_RECIPE = Kitchen::BookRecipe.new(book_short_name: :uphysics) do |doc, resources|
+U_PHYSICS_RECIPE = Kitchen::BookRecipe.new(book_short_name: :uphysics) do |doc, _resources|
   include Kitchen::Directions
 
   book = doc.book
@@ -8,7 +8,6 @@ U_PHYSICS_RECIPE = Kitchen::BookRecipe.new(book_short_name: :uphysics) do |doc, 
 
   book.search('cnx-pi').trash
 
-  BakeImages.v1(book: book, resources: resources)
   BakeUnnumberedFigure.v1(book: book)
   BakePreface.v1(book: book)
   BakeChapterIntroductions.v1(book: book)