+++ spec/recipes_spec/main_spec.rb
@@ -4,229 +4,293 @@ require 'spec_helper'
 
 RSpec.describe 'books' do
   it 'bakes dummy' do
-    expect('dummy').to bake_correctly_with('dummy', '/../books/dummy/resources')
+    cmd = form_bake_cmd(book: 'dummy')
+    expect('dummy').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes accounting' do
-    expect('accounting').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'accounting')
+    expect('accounting').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes american-government' do
-    expect('american-government').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'american-government')
+    expect('american-government').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes anatomy' do
-    expect('anatomy').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'anatomy')
+    expect('anatomy').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes anthropology' do
-    expect('anthropology').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'anthropology')
+    expect('anthropology').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes ap bio' do
-    expect('ap-biology').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'ap-biology')
+    expect('ap-biology').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes ap history' do
-    expect('ap-history').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'ap-history')
+    expect('ap-history').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes ap physics' do
-    expect('ap-physics').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'ap-physics')
+    expect('ap-physics').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes ap physics 2e' do
-    expect('ap-physics-2e').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'ap-physics-2e')
+    expect('ap-physics-2e').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes astronomy' do
-    expect('astronomy').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'astronomy')
+    expect('astronomy').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes bca' do
-    expect('bca').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'bca')
+    expect('bca').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes biology' do
-    expect('biology').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'biology')
+    expect('biology').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes business ethics' do
-    expect('business-ethics').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'business-ethics')
+    expect('business-ethics').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes business law' do
-    expect('business-law').to bake_correctly_with_empty_resources_and_use('business-ethics')
+    cmd = form_bake_cmd(book: 'business-law', recipe: 'business-ethics')
+    expect('business-law').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes calculus' do
-    expect('calculus').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'calculus')
+    expect('calculus').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes chemistry' do
-    expect('chemistry').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'chemistry')
+    expect('chemistry').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes college algebra' do
-    expect('college-algebra').to bake_correctly_with_empty_resources_and_use('precalculus')
+    cmd = form_bake_cmd(book: 'college-algebra', recipe: 'precalculus')
+    expect('college-algebra').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes college algebra coreq' do
-    expect('college-algebra-coreq').to bake_correctly_with_empty_resources_and_use('precalculus-coreq')
+    cmd = form_bake_cmd(book: 'college-algebra-coreq', recipe: 'precalculus-coreq')
+    expect('college-algebra-coreq').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes college physics' do
-    expect('college-physics').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'college-physics')
+    expect('college-physics').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes college physics 2e' do
-    expect('college-physics-2e').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'college-physics-2e')
+    expect('college-physics-2e').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes college success' do
-    expect('college-success').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'college-success')
+    expect('college-success').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes computer science' do
-    expect('computer-science').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'computer-science')
+    expect('computer-science').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes contemporary-math' do
-    expect('contemporary-math').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'contemporary-math')
+    expect('contemporary-math').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes economics' do
-    expect('economics').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'economics')
+    expect('economics').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes elementary algebra' do
-    expect('elementary-algebra').to bake_correctly_with_empty_resources_and_use('dev-math')
+    cmd = form_bake_cmd(book: 'elementary-algebra', recipe: 'dev-math')
+    expect('elementary-algebra').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes english-composition' do
-    expect('english-composition').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'english-composition')
+    expect('english-composition').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes entrepreneurship' do
-    expect('entrepreneurship').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'entrepreneurship')
+    expect('entrepreneurship').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes finance' do
-    expect('finance').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'finance')
+    expect('finance').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes history' do
-    expect('history').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'history')
+    expect('history').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes hs college success' do
-    expect('hs-college-success').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'hs-college-success')
+    expect('hs-college-success').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes hs-physics' do
-    expect('hs-physics').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'hs-physics')
+    expect('hs-physics').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes intellectual property' do
-    expect('intellectual-property').to bake_correctly_with_empty_resources_and_use('business-ethics')
+    cmd = form_bake_cmd(book: 'intellectual-property', recipe: 'business-ethics')
+    expect('intellectual-property').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes intermediate algebra' do
-    expect('intermediate-algebra').to bake_correctly_with_empty_resources_and_use('dev-math')
+    cmd = form_bake_cmd(book: 'intermediate-algebra', recipe: 'dev-math')
+    expect('intermediate-algebra').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes intro business' do
-    expect('intro-business').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'intro-business')
+    expect('intro-business').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes marketing' do
-    expect('marketing').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'marketing')
+    expect('marketing').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes microbio' do
-    expect('microbiology').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'microbiology')
+    expect('microbiology').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes nursing-external' do
-    expect('nursing-external').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'nursing-external')
+    expect('nursing-external').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes nursing-internal' do
-    expect('nursing-internal').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'nursing-internal')
+    expect('nursing-internal').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes organic-chemistry' do
-    expect('organic-chemistry').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'organic-chemistry')
+    expect('organic-chemistry').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes organizational-behavior' do
-    expect('organizational-behavior').to bake_correctly_with_empty_resources_and_use('principles-management')
+    cmd = form_bake_cmd(book: 'organizational-behavior', recipe: 'principles-management')
+    expect('organizational-behavior').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes philosophy' do
-    expect('philosophy').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'philosophy')
+    expect('philosophy').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes pl-marketing' do
-    expect('pl-marketing').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'pl-marketing')
+    expect('pl-marketing').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes pl-microeconomics' do
-    expect('pl-microeconomics').to bake_correctly_with_empty_resources_and_use('pl-economics')
+    cmd = form_bake_cmd(book: 'pl-microeconomics', recipe: 'pl-economics')
+    expect('pl-microeconomics').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes pl-psychology' do
-    expect('pl-psychology').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'pl-psychology')
+    expect('pl-psychology').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes pl u-physics' do
-    expect('pl-u-physics').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'pl-u-physics')
+    expect('pl-u-physics').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes political science' do
-    expect('political-science').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'political-science')
+    expect('political-science').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes prealgebra' do
-    expect('prealgebra').to bake_correctly_with_empty_resources_and_use('dev-math')
+    cmd = form_bake_cmd(book: 'prealgebra', recipe: 'dev-math')
+    expect('prealgebra').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes precalculus' do
-    expect('precalculus').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'precalculus')
+    expect('precalculus').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes principles-management' do
-    expect('principles-management').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'principles-management')
+    expect('principles-management').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes psychology' do
-    expect('psychology').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'psychology')
+    expect('psychology').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes python' do
-    expect('python').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'python')
+    expect('python').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes sociology' do
-    expect('sociology').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'sociology')
+    expect('sociology').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes statistics' do
-    expect('statistics').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'statistics')
+    expect('statistics').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes trigonometry' do
-    expect('trigonometry').to bake_correctly_with_empty_resources_and_use('precalculus')
+    cmd = form_bake_cmd(book: 'trigonometry', recipe: 'precalculus')
+    expect('trigonometry').to match_expected_when_baked_with(cmd)
   end
 
   it 'bakes u-physics' do
-    expect('u-physics').to bake_correctly_with_empty_resources
+    cmd = form_bake_cmd(book: 'u-physics')
+    expect('u-physics').to match_expected_when_baked_with(cmd)
   end
 
-  it 'bakes world-histrory' do
-    expect('world-history').to bake_correctly_with_empty_resources
+  it 'bakes world-history' do
+    cmd = form_bake_cmd(book: 'world-history')
+    expect('world-history').to match_expected_when_baked_with(cmd)
   end
 
   # ATTENTION: SPECS ARE NOW ALPHABETIZED
   # please add future specs in alphabetical order
+
+  it 'bakes with the web pipeline on test data' do
+    cmd = form_bake_cmd(
+      book: 'web-test', recipe: 'dummy',
+      resource_path: '../books/web-test/resources', output_platform: 'web'
+    )
+    expect('web-test').to match_expected_when_baked_with(cmd)
+  end
 end